feat: 기사 뷰 등록 테스트 그린 단계 작성

+++ src/main/java/com/example/monew/domain/article/dto/ArticleViewDto.java
@@ -14,6 +14,6 @@ public record ArticleViewDto(
         String articleTitle,
         LocalDateTime articlePublishedDate,
         String articleSummary,
-        Integer articleCommentCount,
-        Integer articleViewCount
+        long articleCommentCount,
+        long articleViewCount
 ) {}
